Add WinDrawer class to encapsulate windows' state and actions
The current code has different dictionaries for each class of
interaction with the windows (modifying their state, passing no
parameters to the functions, just one, etc.). This clearly doesn't
scale, so an object will keep the window objects and will interact
with them, hiding the details from the main function.